torchserve/realesrgan/requirements.txt
accelerate==0.15.0
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
enum-compat==0.0.3
executing==1.2.0
ffmpeg-python==0.2.0
filelock==3.9.0
fonttools==4.38.0
future==0.18.3
huggingface-hub==0.12.0
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.20.2
ipython==8.8.0
jedi==0.18.2
jupyter_client==8.0.1
jupyter_core==5.1.5
kiwisolver==1.4.4
lpips==0.1.4
matplotlib==3.6.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.1
opencv-python==4.7.0.68
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pyiqa==0.1.5
piqa==1.2.2
platformdirs==2.6.2
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==0.21.1
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
timm==0.6.12
torch-model-archiver==0.7.0
torchserve==0.7.0
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
wget==3.2
yacs==0.1.8
zipp==3.11.0
nvgpu==0.9.0
pynvml==11.5.0
